tests/requirements-test.txt
# Test dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-asyncio>=0.21.1

# Service dependencies needed for tests
influxdb-client>=1.36.0
speedtest-cli>=2.1.3
python-wifi>=0.6.1
docker>=6.1.3
pyyaml>=6.0.1
schedule>=1.2.0
tenacity>=8.2.2
